CsvManager (csv_manager v0.1.0)

CsvManager is a csv manipulator built with and made for elixir programs.

Functions

csv_info(data)

Display csv info. Inputs %CsvManager.Data(the data structure)

get_column(data, column_name)

Acess all the values from a column Inputs data and column_name

get_row(data, number)

Returns a specific row from the data Inputs data and number (number of the row)

map_columns(data)

Function to create map with columns-> column_index

parse_data(data)

Function responsible for parsing the string into a list of lists.

read_csv(filename)

Main function to read a csv file. Input filename(the file name) the module estipulates a "," for spliting rows

read_csv(filename, sep)

Main function to read a csv file. Inputs filename(the file name) and sep(character to split the rows)

read_file(filename, sep)

Function that reads the file by a filename. Inputs: filename (name of the file), sep (character that separates columns).

Example

iex> CsvManager.read_file("example.csv", ";")
"The example.csv couldn't be read by: enoent"